requirements.txt
werkzeug
Flask==1.1.1
Flask-Cors==3.0.8
Pillow==6.2.1
pdf2image==1.9.0
wheel==0.33.6
pytest==5.2.2
pytest-cov==2.8.1
tesserocr==2.4.1
